refactor: restaurantList 유틸 관련 로직 분리
